Evm upgrades and tests (#38)
* Add event testing to icco.js

* Add testnet upgrade script

* Add wormhole fee to tests

* Generate Solana specific VAAs in Conductor.sol

* Add solana contract to testnet.json

+++ ethereum/contracts/icco/conductor/ConductorGetters.sol
@@ -39,11 +39,11 @@ contract ConductorGetters is ConductorState {
         return _state.contributorImplementations[chainId_];
     }
 
-    function solanaWallet(uint saleId_) public view returns (bytes32) {
+    function solanaWallet(uint256 saleId_) public view returns (bytes32) {
         return _state.sales[saleId_].solanaTokenAccount;
     }
 
-    function contributorWallets(uint saleId_, uint16 chainId_) public view returns (bytes32) {
+    function contributorWallets(uint256 saleId_, uint16 chainId_) public view returns (bytes32) {
         /// @dev Solana chainID == 1
         if (chainId_ == 1) {
             return solanaWallet(saleId_); 
@@ -52,19 +52,19 @@ contract ConductorGetters is ConductorState {
         }
     }
 
-    function sales(uint saleId_) public view returns (ConductorStructs.Sale memory sale) {
+    function sales(uint256 saleId_) public view returns (ConductorStructs.Sale memory sale) {
         return _state.sales[saleId_];
     }
 
-    function getNextSaleId() public view returns (uint) {
+    function getNextSaleId() public view returns (uint256) {
         return _state.nextSaleId;
     }
 
-    function saleContributionIsCollected(uint saleId_, uint tokenIndex) public view returns (bool) {
+    function saleContributionIsCollected(uint256 saleId_, uint256 tokenIndex) public view returns (bool) {
         return _state.sales[saleId_].contributionsCollected[tokenIndex];
     }
 
-    function saleContributions(uint saleId_) public view returns (uint[] memory) {
+    function saleContributions(uint256 saleId_) public view returns (uint256[] memory) {
         return _state.sales[saleId_].contributions;
     }
 }
